+++ Changelog.md
@@ -1,4 +1,4 @@
-# Dev Changelog
+# Changelog
 
 ### v0.1.5