+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v0.1.6
+
+- \[BREAKING CHANGE\] all components custom style props should pass to props.customize
+- \[BREAKING CHANGE\] component node class name will change to `UUI-XXX-YYY`
+- Storybook add Notes story, Storybook support code highlight
+
 ### v0.1.5
 
 - export `Label`, `Pagination`, `Select` and `Toast` components